E2E/Playwright: Group related batch 5 specs
